+++ crates/light-client-prover/src/lib.rs
@@ -19,6 +19,10 @@ pub mod da_block_handler;
 #[cfg(feature = "native")]
 pub mod db_migrations;
 #[cfg(feature = "native")]
+mod input_builder;
+#[cfg(feature = "native")]
+mod lcp_storage;
+#[cfg(feature = "native")]
 pub mod metrics;
 #[cfg(feature = "native")]
 pub mod rpc;